refactor: merged booting, tupos, resolver, update logging, deps,  fixed locks, simplify api, optimisations.

+++ src/webdav/WebdavFileSystem.ts
@@ -1,13 +1,16 @@
 import { Readable, Writable } from "stream";
 import { ResourceType, v2, Errors } from "webdav-server";
-import { LocalLockManager, LocalPropertyManager } from "webdav-server/lib/index.v2.js";
+import { LocalLockManager, LocalPropertyManager, Resource } from "webdav-server/lib/index.v2.js";
 
 
 import mime from "mime-types";
 import path from "path";
+import { createVFile } from "../file/IFile";
+
 import DICloudApp from "../DICloudApp.js";
 import VolumeEx from "../file/VolumeEx.js";
 import Log from "../Log.js";
+import { ENCRYPTION_OVERHEAD } from "../provider/DiscordFileProvider";
 
 
 class VirtualDiscordFileSystemSerializer implements v2.FileSystemSerializer {
@@ -40,10 +43,19 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
 
     protected _propertyManager(path: v2.Path, ctx: v2.PropertyManagerInfo, callback: v2.ReturnCallback<v2.IPropertyManager>): void {
         const key = path.toString();
+        if (key == "/.git/info/exclude"){
+            Log.info("test")
+        }
+
         if (!this.properties.has(key)) {
             this.properties.set(key, new LocalPropertyManager());
         }
         return callback(undefined, this.properties.get(key));
+    }
+
+    private cleanupLocksAndProperties(path: v2.Path) {
+        this.locks.delete(path.toString());
+        this.properties.delete(path.toString());
     }
 
     /**
@@ -62,7 +74,12 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
         const stat = this.fs.statSync(path.toString());
 
         if (stat.isFile()) {
-            return callback(undefined, this.fs.getFile(path.toString()).size);
+            const file = this.fs.getFile(path.toString());
+
+            if(file.encrypted){
+                return callback(undefined, file.size - (ENCRYPTION_OVERHEAD * file.chunks.length)); // -16 bytes for each chunk for encryption metadata. client side will wait for full file, so if we provide size with metadata, which exists only on the server, the client will wait for the metadata to be downloaded which will never happen.
+            }
+            return callback(undefined, file.size);
         }
         return callback(undefined, this.fs.getTreeSizeRecursive(path.toString()));
     }
@@ -80,7 +97,7 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
     }
 
     protected _type(path: v2.Path, ctx: v2.TypeInfo, callback: v2.ReturnCallback<v2.ResourceType>): void {
-        Log.info(".type", path.toString(), getContext(ctx));
+        // Log.info(".type", path.toString(), getContext(ctx));
 
         const stat = this.fs.statSync(path.toString());
 
@@ -90,7 +107,7 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
             return callback(undefined, ResourceType.Directory);
         }
 
-        return callback(Errors.ResourceNotFound);
+        return callback(undefined, ResourceType.NoResource);
     }
 
     protected _mimeType(path: v2.Path, ctx: v2.MimeTypeInfo, callback: v2.ReturnCallback<string>): void {
@@ -107,7 +124,6 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
 
     protected _fastExistCheck(ctx: v2.RequestContext, path: v2.Path, callback: (exists: boolean) => void): void {
         // Log.info(".fastExistCheck", path.toString(), getContext(ctx));
-        Log.info(".fastExistCheck", path.toString());
 
         return callback(this.fs.existsSync(path.toString()));
     }
@@ -120,7 +136,7 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
         }
 
         if (ctx.type.isFile) {
-            this.fs.setFile(path.toString(), this.client.getProvider().createVFile(path.fileName(), 0));
+            this.fs.setFile(path.toString(), createVFile(path.fileName(), 0, this.client.shouldEncryptFiles()));
         }
 
         this.client.markForUpload();
@@ -134,25 +150,26 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
 
         const stat = this.fs.statSync(path.toString());
 
-        if (!stat.isFile()) {
+        if (!stat.isFile()){
             return callback(Errors.ResourceNotFound);
         }
 
         const file = this.fs.getFile(path.toString())
-        if (file.chunks.length == 0) {
-            return callback(undefined, Readable.from(Buffer.from([])));
+        if (file.size == 0 || file.chunks.length == 0) {
+            Log.info(".openReadStream", "File has no chunks: " + path.toString());
+            return callback(undefined, Readable.from([]));
         }
 
-        Log.info(".openReadStream, fetching: ", file);
+        Log.info(".openReadStream, creating read stream: ", file);
         const readStream = await this.client.getProvider().createReadStream(file)
         Log.info(".openReadStream", "Stream opened: " + path.toString());
-
         return callback(undefined, readStream);
     }
 
+
     async _openWriteStream(path: v2.Path, ctx: v2.OpenWriteStreamInfo, callback: v2.ReturnCallback<Writable>): Promise<void> {
         const { targetSource, estimatedSize, mode } = ctx;
-        Log.info(".openWriteStream", targetSource, estimatedSize, mode, "shouldEncrypt: ", this.client.shouldEncryptFiles());
+        Log.info(".openWriteStream", targetSource, estimatedSize, mode);
 
         const stat = this.fs.statSync(path.toString());
 
@@ -162,6 +179,7 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
 
         const file = this.fs.getFile(path.toString());
 
+        // overwrite file
         this.client.getProvider().addToDeletionQueue(file.chunks.map(chunk => ({
             channel: this.client.getFilesChannel().id,
             message: chunk.id
@@ -194,6 +212,10 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
     async _delete(path: v2.Path, ctx: v2.DeleteInfo, callback: v2.SimpleCallback): Promise<void> {
         Log.info(".delete", path.toString(), getContext(ctx));
 
+        if (path.toString() == "/") {
+            return callback(Errors.InvalidOperation);
+        }
+
         const stat = this.fs.statSync(path.toString());
         const filesToDelete: string[] = [];
 
@@ -205,8 +227,6 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
             filesToDelete.push(...this.fs.getPathsRecursive(path.toString()));
         }
 
-        console.dir(filesToDelete);
-
         for (const fileToDelete of filesToDelete) {
             for (const chunk of this.fs.getFile(fileToDelete).chunks) {
                 this.client.getProvider().addToDeletionQueue([{
@@ -216,9 +236,15 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
             }
         }
 
-
         this.fs.rmSync(path.toString(), { recursive: true });
         this.client.markForUpload();
+
+        for (const fileToDelete of filesToDelete) {
+            this.cleanupLocksAndProperties(new v2.Path(fileToDelete));
+        }
+        this.cleanupLocksAndProperties(path);
+
+
         return callback();
     }
 
@@ -237,7 +263,7 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
             this.fs.mkdirSync(path.parse(pathTo.toString()).dir, { recursive: true });
 
             const oldFile = this.fs.getFile(pathFrom.toString());
-            const newFile = this.client.getProvider().createVFile(pathTo.fileName(), oldFile.size);
+            const newFile = createVFile(pathTo.fileName(), oldFile.size, oldFile.encrypted);
 
             const readStream = await this.client.getProvider().createReadStream(oldFile);
             const writeStream = await this.client.getProvider().createWriteStream(newFile);
@@ -250,8 +276,16 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
             writeStream.on("finish", () => {
                 Log.info(".copy", "Stream finished: " + pathTo.toString());
                 this.fs.setFile(pathTo.toString(), newFile);
-                this.client.markForUpload();
+                
+                this.locks.set(pathTo.toString(), new LocalLockManager());
+                const oldProps = this.properties.get(pathFrom.toString());
+                if (oldProps) {
+                    this.properties.set(pathTo.toString(), oldProps);
+                }
 
+                
+                this.client.markForUpload();
+                
                 return resolve(true);
             });
 
@@ -297,12 +331,29 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
         const sourceExists = this.fs.existsSync(pathFrom.toString());
         const targetExists = this.fs.existsSync(pathTo.toString());
 
-        if (!sourceExists || targetExists) {
+        if (!sourceExists || (targetExists && !ctx.overwrite)) {
             return callback(Errors.InvalidOperation);
         }
 
+        
+        if(ctx.overwrite && targetExists){
+            const targetStat = this.fs.statSync(pathTo.toString());
+            if(targetStat.isFile()){
+                this.fs.rmSync(pathTo.toString(), { force: true });
+            } else {
+                return callback(Errors.InvalidOperation);
+            }
+        }
+        
         this.fs.renameSync(pathFrom.toString(), pathTo.toString());
         this.client.markForUpload();
+
+        this.locks.set(pathTo.toString(), new LocalLockManager());
+        const oldProps = this.properties.get(pathFrom.toString());
+        if (oldProps) {
+            this.properties.set(pathTo.toString(), oldProps);
+        }
+        this.cleanupLocksAndProperties(pathFrom);
 
         return callback(undefined, true);
     }
@@ -313,7 +364,25 @@ export default class DiscordWebdavFilesystemHandler extends v2.FileSystem {
         const oldPath = pathFrom.toString();
         const newPath = pathFrom.parentName() + "/" + newName;
 
+        if (!this.fs.existsSync(oldPath)) {
+            return callback(Errors.ResourceNotFound);
+        }
+
+        if (this.fs.existsSync(newPath)) {
+            return callback(Errors.ResourceAlreadyExists);
+        }
+
+        this.locks.set(newPath, new LocalLockManager());
+        const oldProps = this.properties.get(oldPath);
+        if (oldProps) {
+            this.properties.set(newPath, oldProps);
+        }
+        this.cleanupLocksAndProperties(pathFrom);
+
+
         this.fs.renameSync(oldPath, newPath);
+        this.client.markForUpload();
+
         callback(undefined, true);
     }
